Async-ready resource system added WPF project.
New resource system layout that is async ready. It follows the Unreal
model of "importing" and asset, thus converting it to an engine-ready
format and saving that as binary data. Asset loading can be from a .bin
file or a project structure. Added an editor WPF project that will be the
frontend for the engine.