Second stage of the milti-threaded rendering pipeline.
The second piece of a fully multi-threaded rendering pipeline: resource
loading on the GPU thread. Just like before, all resource pre-processing
is done off the GPU thread before being marshaled over to the GPU thread.
Data is, for the most part, in a ready-to-upload state by the time the GPU
thread sees it.