Add HTML Flex style GUI layout, reworked support code for that.
Add in the bases for HTML Flex style Gui layout, and reworked a lot of
code to support multiple pipeline passes for a single target and other
things needed for GUI rendering. Need to make GLTexture use double
buffering next, as the GC load is unacceptable with all these Bitmap
copies being made.